extract notifyPrefs for both user and organisation
