+++ backend/lib/endpoints/schema/organisations.js
@@ -1,6 +1,7 @@
 const S = require("fluent-schema");
 const { strictSchema } = require("./utils");
 const { locationSchema } = require("./location");
+const { notifyPreferenceSchema } = require("./notificationPreference");
 
 const organisation = {
   about: S.string().maxLength(100),
@@ -37,6 +38,7 @@ const createOrganisationSchema = {
     .prop("needs", organisation.needs)
     .prop("type", organisation.type.required())
     .prop("urls", organisation.urls)
+    .prop("notifyPrefs", notifyPreferenceSchema)
     .required(["location"]),
 };
 
@@ -78,7 +80,8 @@ const updateOrganisationSchema = {
     .prop("name", organisation.name)
     .prop("needs", organisation.needs)
     .prop("type", organisation.type)
-    .prop("urls", organisation.urls),
+    .prop("urls", organisation.urls)
+    .prop("notifyPrefs", notifyPreferenceSchema),
   params: strictSchema().prop("organisationId", S.string().required()),
 };
